requirements.txt
anyio==3.7.0
attrs==22.1.0
automata-lib==7.1.0
build==0.10.0
CacheControl==0.12.11
certifai==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.21.3
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
networkx==3.1
packaging==22.0
pexpect==4.8.0
pipenv==2023.4.29
pkginfo==1.9.6
platformdirs==2.5.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
pyparsing==3.1.0
pyproject_hooks==1.0.0
pyrsistent==0.18.1
python-dotenv==1.0.0
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.6.3
urllib3==1.26.12
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
